+++ modules/core/modelList/layer/vectorBase.js
@@ -89,9 +89,7 @@ const VectorBaseLayer = Layer.extend(/** @lends GeoJSONLayer.prototype */{
         }
 
         if (styleModel && legend === true) {
-            if (Config.hasOwnProperty("useVectorStyleBeta") && Config.useVectorStyleBeta ? Config.useVectorStyleBeta : false) {
-                styleModel.getGeometryTypeFromWFS(this.get("url"), this.get("version"), this.get("featureType"), this.get("styleGeometryType"));
-            }
+            styleModel.getGeometryTypeFromWFS(this.get("url"), this.get("version"), this.get("featureType"), this.get("styleGeometryType"));
             this.setLegend(styleModel.getLegendInfos());
         }
         else if (typeof legend === "string") {