+++ portal/masterDefault/config.js
@@ -121,13 +121,7 @@ const Config = {
         fallbackLanguage: "de",
         changeLanguageOnStartWhen: ["querystring", "localStorage", "navigator", "htmlTag"],
         loadPath: "/locales/{{lng}}/{{ns}}.json"
-    },
-    /**
-     * @deprecated to be deleted with version 3.0
-     * @property {boolean} Config.useVectorStyleBeta Flag to use the new vectorStyling module for backward compatibility
-     * @default false
-     */
-    useVectorStyleBeta: true
+    }
 };
 
 // conditional export to make config readable by e2e tests